Move mapping file
